fixed image support for post_db_repo

+++ services/image_repo.py
@@ -1,15 +1,16 @@
 import os
 
-from interfaces.image_support_interface import IImageRepo
+from interfaces.image_repo_interface import IImageRepo
 
 
 class ImageRepo(IImageRepo):
     def remove_image(self, filename):
-        path = './static/images/' + filename
-        if os.path.exists(path):
-            os.remove(path)
-        else:
-            print('The file doesnt exists')
+        if filename != '' and filename is not None:
+            path = './static/images/' + filename
+            if os.path.exists(path):
+                os.remove(path)
+            else:
+                print('The file doesnt exists')
 
     def save_image(self, file, filename):
         if file:
